Move cases to separate source file, use tags and prefixes instead of objects to group cases.
